Move kazoo package
